fix: blockchain listener
