+++ apps/api/package.json
@@ -37,7 +37,8 @@
     "ioredis": "^5.10.1",
     "lodash": "^4.17.23",
     "reflect-metadata": "^0.2.2",
-    "rxjs": "^7.8.1"
+    "rxjs": "^7.8.1",
+    "type-fest": "^5.5.0"
   },
   "devDependencies": {
     "@eslint/eslintrc": "^3.2.0",
