--- apps/api/src/loan/dto/update-loan.dto.ts
@@ -1,4 +1,0 @@
-import { PartialType } from '@nestjs/mapped-types';
-import { CreateLoanDto } from './create-loan.dto';
-
-export class UpdateLoanDto extends PartialType(CreateLoanDto) {}
