--- apps/api/src/loan/entities/loan.entity.ts
@@ -1,1 +1,0 @@
-export class Loan {}
