+++ apps/api/src/loan/loan.module.ts
@@ -1,11 +1,10 @@
 import { Module } from '@nestjs/common';
 import { SupabaseModule } from '../supabase/supabase.module';
-import { LoanController } from './loan.controller';
 import { LoanService } from './loan.service';
 
 @Module({
   imports: [SupabaseModule],
-  controllers: [LoanController],
   providers: [LoanService],
+  exports: [LoanService],
 })
 export class LoanModule {}
