+++ apps/api/src/loan/loan.service.ts
@@ -6,14 +6,23 @@ import { CreateLoanDto } from './dto/create-loan.dto';
 export class LoanService {
   constructor(private readonly supabaseService: SupabaseService) {}
 
-  async create(createLoanDto: CreateLoanDto, borrower: string) {
-    const { data, error } = await this.supabaseService.client
+  async create({ collateralTokenAddress, ...createLoanDto }: CreateLoanDto) {
+    const { data: token, error: tokenError } = await this.supabaseService.client
+      .from('token_list')
+      .select('id::text')
+      .eq('address', collateralTokenAddress)
+      .eq('chainId', createLoanDto.chainId)
+      .single();
+
+    if (tokenError || !token)
+      throw new Error(
+        `Collateral token not found in token_list: ${collateralTokenAddress} on chain ${createLoanDto.chainId}`,
+      );
+
+    const { error } = await this.supabaseService.client
       .from('loans')
-      .insert({ borrower, ...createLoanDto })
-      .select('*');
+      .insert({ ...createLoanDto, collateralTokenId: token.id });
 
     if (error) throw error;
-
-    return data?.[0];
   }
 }
