+++ apps/api/src/supabase/database.types.ts
@@ -1,263 +1,34 @@
-export type Json =
-  | string
-  | number
-  | boolean
-  | null
-  | { [key: string]: Json | undefined }
-  | Json[]
+import { MergeDeep } from 'type-fest';
+import { Database as DatabaseGenerated } from './database-generated.types';
 
-export type Database = {
-  graphql_public: {
-    Tables: {
-      [_ in never]: never
-    }
-    Views: {
-      [_ in never]: never
-    }
-    Functions: {
-      graphql: {
-        Args: {
-          extensions?: Json
-          operationName?: string
-          query?: string
-          variables?: Json
-        }
-        Returns: Json
-      }
-    }
-    Enums: {
-      [_ in never]: never
-    }
-    CompositeTypes: {
-      [_ in never]: never
-    }
+export type Database = MergeDeep<
+  DatabaseGenerated,
+  {
+    public: {
+      Tables: {
+        loans: {
+          Row: {
+            loanId: string;
+            chainId: string;
+            collateralAmount: string;
+            collateralBlockNumber: string;
+          };
+          Insert: {
+            loanId: string;
+            chainId: string;
+            collateralAmount: string;
+            collateralBlockNumber: string;
+          };
+        };
+        token_list: {
+          Row: {
+            chainId: string;
+          };
+          Insert: {
+            chainId: string;
+          };
+        };
+      };
+    };
   }
-  public: {
-    Tables: {
-      loans: {
-        Row: {
-          borrower: string
-          chainId: number
-          collateralAmount: number
-          collateralBlockHash: string | null
-          collateralBlockNumber: number | null
-          collateralLockedAt: string | null
-          collateralTokenId: string
-          collateralTxHash: string | null
-          createdAt: string
-          id: string
-          status: Database["public"]["Enums"]["loan_status"]
-        }
-        Insert: {
-          borrower: string
-          chainId?: number
-          collateralAmount: number
-          collateralBlockHash?: string | null
-          collateralBlockNumber?: number | null
-          collateralLockedAt?: string | null
-          collateralTokenId: string
-          collateralTxHash?: string | null
-          createdAt?: string
-          id?: string
-          status?: Database["public"]["Enums"]["loan_status"]
-        }
-        Update: {
-          borrower?: string
-          chainId?: number
-          collateralAmount?: number
-          collateralBlockHash?: string | null
-          collateralBlockNumber?: number | null
-          collateralLockedAt?: string | null
-          collateralTokenId?: string
-          collateralTxHash?: string | null
-          createdAt?: string
-          id?: string
-          status?: Database["public"]["Enums"]["loan_status"]
-        }
-        Relationships: [
-          {
-            foreignKeyName: "loans_collateralTokenId_fkey"
-            columns: ["collateralTokenId"]
-            isOneToOne: false
-            referencedRelation: "token_list"
-            referencedColumns: ["id"]
-          },
-        ]
-      }
-      token_list: {
-        Row: {
-          address: string
-          chainId: number
-          decimals: number | null
-          id: string
-          logoURI: string | null
-          name: string | null
-          symbol: string
-          updatedAt: string
-        }
-        Insert: {
-          address: string
-          chainId: number
-          decimals?: number | null
-          id?: string
-          logoURI?: string | null
-          name?: string | null
-          symbol: string
-          updatedAt?: string
-        }
-        Update: {
-          address?: string
-          chainId?: number
-          decimals?: number | null
-          id?: string
-          logoURI?: string | null
-          name?: string | null
-          symbol?: string
-          updatedAt?: string
-        }
-        Relationships: []
-      }
-    }
-    Views: {
-      [_ in never]: never
-    }
-    Functions: {
-      [_ in never]: never
-    }
-    Enums: {
-      loan_status: "pending" | "active" | "repaid" | "defaulted" | "cancelled"
-    }
-    CompositeTypes: {
-      [_ in never]: never
-    }
-  }
-}
-
-type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">
-
-type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">]
-
-export type Tables<
-  DefaultSchemaTableNameOrOptions extends
-    | keyof (DefaultSchema["Tables"] & DefaultSchema["Views"])
-    | { schema: keyof DatabaseWithoutInternals },
-  TableName extends DefaultSchemaTableNameOrOptions extends {
-    schema: keyof DatabaseWithoutInternals
-  }
-    ? keyof (DatabaseWithoutInternals[DefaultSchemaTableNameOrOptions["schema"]]["Tables"] &
-        DatabaseWithoutInternals[DefaultSchemaTableNameOrOptions["schema"]]["Views"])
-    : never = never,
-> = DefaultSchemaTableNameOrOptions extends {
-  schema: keyof DatabaseWithoutInternals
-}
-  ? (DatabaseWithoutInternals[DefaultSchemaTableNameOrOptions["schema"]]["Tables"] &
-      DatabaseWithoutInternals[DefaultSchemaTableNameOrOptions["schema"]]["Views"])[TableName] extends {
-      Row: infer R
-    }
-    ? R
-    : never
-  : DefaultSchemaTableNameOrOptions extends keyof (DefaultSchema["Tables"] &
-        DefaultSchema["Views"])
-    ? (DefaultSchema["Tables"] &
-        DefaultSchema["Views"])[DefaultSchemaTableNameOrOptions] extends {
-        Row: infer R
-      }
-      ? R
-      : never
-    : never
-
-export type TablesInsert<
-  DefaultSchemaTableNameOrOptions extends
-    | keyof DefaultSchema["Tables"]
-    | { schema: keyof DatabaseWithoutInternals },
-  TableName extends DefaultSchemaTableNameOrOptions extends {
-    schema: keyof DatabaseWithoutInternals
-  }
-    ? keyof DatabaseWithoutInternals[DefaultSchemaTableNameOrOptions["schema"]]["Tables"]
-    : never = never,
-> = DefaultSchemaTableNameOrOptions extends {
-  schema: keyof DatabaseWithoutInternals
-}
-  ? DatabaseWithoutInternals[DefaultSchemaTableNameOrOptions["schema"]]["Tables"][TableName] extends {
-      Insert: infer I
-    }
-    ? I
-    : never
-  : DefaultSchemaTableNameOrOptions extends keyof DefaultSchema["Tables"]
-    ? DefaultSchema["Tables"][DefaultSchemaTableNameOrOptions] extends {
-        Insert: infer I
-      }
-      ? I
-      : never
-    : never
-
-export type TablesUpdate<
-  DefaultSchemaTableNameOrOptions extends
-    | keyof DefaultSchema["Tables"]
-    | { schema: keyof DatabaseWithoutInternals },
-  TableName extends DefaultSchemaTableNameOrOptions extends {
-    schema: keyof DatabaseWithoutInternals
-  }
-    ? keyof DatabaseWithoutInternals[DefaultSchemaTableNameOrOptions["schema"]]["Tables"]
-    : never = never,
-> = DefaultSchemaTableNameOrOptions extends {
-  schema: keyof DatabaseWithoutInternals
-}
-  ? DatabaseWithoutInternals[DefaultSchemaTableNameOrOptions["schema"]]["Tables"][TableName] extends {
-      Update: infer U
-    }
-    ? U
-    : never
-  : DefaultSchemaTableNameOrOptions extends keyof DefaultSchema["Tables"]
-    ? DefaultSchema["Tables"][DefaultSchemaTableNameOrOptions] extends {
-        Update: infer U
-      }
-      ? U
-      : never
-    : never
-
-export type Enums<
-  DefaultSchemaEnumNameOrOptions extends
-    | keyof DefaultSchema["Enums"]
-    | { schema: keyof DatabaseWithoutInternals },
-  EnumName extends DefaultSchemaEnumNameOrOptions extends {
-    schema: keyof DatabaseWithoutInternals
-  }
-    ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
-    : never = never,
-> = DefaultSchemaEnumNameOrOptions extends {
-  schema: keyof DatabaseWithoutInternals
-}
-  ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
-  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
-    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
-    : never
-
-export type CompositeTypes<
-  PublicCompositeTypeNameOrOptions extends
-    | keyof DefaultSchema["CompositeTypes"]
-    | { schema: keyof DatabaseWithoutInternals },
-  CompositeTypeName extends PublicCompositeTypeNameOrOptions extends {
-    schema: keyof DatabaseWithoutInternals
-  }
-    ? keyof DatabaseWithoutInternals[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"]
-    : never = never,
-> = PublicCompositeTypeNameOrOptions extends {
-  schema: keyof DatabaseWithoutInternals
-}
-  ? DatabaseWithoutInternals[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"][CompositeTypeName]
-  : PublicCompositeTypeNameOrOptions extends keyof DefaultSchema["CompositeTypes"]
-    ? DefaultSchema["CompositeTypes"][PublicCompositeTypeNameOrOptions]
-    : never
-
-export const Constants = {
-  graphql_public: {
-    Enums: {},
-  },
-  public: {
-    Enums: {
-      loan_status: ["pending", "active", "repaid", "defaulted", "cancelled"],
-    },
-  },
-} as const
-
+>;
