+++ apps/api/src/token-list/token-list.mock.ts
@@ -2,26 +2,24 @@ import { TokenListResponse } from './token-list.service';
 
 export const tokenListMock = (
   HARDCODED_MOCK_ERC20_ADDRESS: string,
-): TokenListResponse => ({
-  tokens: {
-    '1337': [
-      {
-        chainId: 1337,
-        address: '0x0000000000000000000000000000000000000000',
-        symbol: 'ETH',
-        name: 'ETH',
-        decimals: 18,
-        logoURI:
-          'https://raw.githubusercontent.com/trustwallet/assets/master/blockchains/ethereum/assets/0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2/logo.png',
-      },
-      {
-        chainId: 1337,
-        address: HARDCODED_MOCK_ERC20_ADDRESS,
-        symbol: 'MOCK',
-        name: 'MockToken',
-        decimals: 18,
-        logoURI: null,
-      },
-    ],
-  },
+): TokenListResponse['tokens'] => ({
+  '1337': [
+    {
+      chainId: 1337,
+      address: '0x0000000000000000000000000000000000000000',
+      symbol: 'ETH',
+      name: 'ETH',
+      decimals: 18,
+      logoURI:
+        'https://raw.githubusercontent.com/trustwallet/assets/master/blockchains/ethereum/assets/0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2/logo.png',
+    },
+    {
+      chainId: 1337,
+      address: HARDCODED_MOCK_ERC20_ADDRESS,
+      symbol: 'MOCK',
+      name: 'MockToken',
+      decimals: 18,
+      logoURI: null,
+    },
+  ],
 });
